+++ app/models/computation.rb
@@ -12,6 +12,8 @@ class Computation < ApplicationRecord
 
   scope :active, -> { where(status: %w[new queued running]) }
   scope :submitted, -> { where(status: %w[queued running]) }
+  scope :created, -> { where(status: 'created') }
+  scope :not_finished, -> { where(status: %w[created new queued running]) }
   scope :rimrock, -> { where(type: 'RimrockComputation') }
   scope :webdav, -> { where(type: 'WebdavComputation') }
   scope :submitted_rimrock, -> { submitted.rimrock }
@@ -19,6 +21,7 @@ class Computation < ApplicationRecord
   scope :for_patient_status, ->(status) { where(pipeline_step: status) }
 
   delegate :runnable?, :run, to: :runner
+  delegate :mode, :manual?, :automatic?, to: :pipeline
 
   def active?
     %w[new queued running].include? status