+++ app/models/pipeline.rb
@@ -10,6 +10,8 @@ class Pipeline < ApplicationRecord
 
   FLOWS = %w[full_body_scan partial_body_scan something_else].freeze
 
+  enum mode: [:automatic, :manual]
+
   belongs_to :patient
   belongs_to :user
   has_many :data_files
@@ -18,6 +20,9 @@ class Pipeline < ApplicationRecord
   validate :set_iid, on: :create
   validates :iid, presence: true, numericality: true
   validates :name, presence: true
+  validates :mode, presence: true
+
+  scope :automatic, -> { where(mode: :automatic) }
 
   validates :flow,
             inclusion: { in: Pipeline::FLOWS }