+++ app/policies/rimrock_computation_policy.rb
@@ -4,4 +4,17 @@ class RimrockComputationPolicy < ApplicationPolicy
   def permitted_attributes
     [:tag_or_branch]
   end
+
+  def show?
+    true
+  end
+
+  def update?
+    record.user == user &&
+      (record.manual? ? record.runnable? : record.tag_or_branch.nil?)
+  end
+
+  def need_proxy?
+    !(record.finished? || Proxy.new(user)&.valid?)
+  end
 end