+++ app/services/pipeline_updater.rb
@@ -9,5 +9,6 @@ class PipelineUpdater
     patient = @computation.pipeline.patient
     user = @computation.user
     patient && user && patient.execute_data_sync(user)
+    Pipelines::StartRunnable.new(@computation.pipeline).call
   end
 end