+++ db/migrate/20170828113755_add_mode_to_pipeline.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class AddModeToPipeline < ActiveRecord::Migration[5.1]
+  def change
+    add_column :pipelines, :mode, :integer, default: 0, null: false
+  end
+end