+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 ### Added
 - Show started rimrock computation source link (@mkasztelnik)
 - Parameter extraction pipeline step (@amber7b)
+- Automatic and manual pipeline execution mode (@mkasztelnik)
 
 ### Changed
 - Segmentation run mode can be configured using yaml or ENV variable (@mkasztelnik)